# api/openapi.yaml
openapi: 3.0.3
info:
  title: SukumadPro API
  version: 0.1.0
  description: |
    OpenAPI reference for the Gin HTTP API used by the web and desktop clients.

    This specification is intentionally aligned to the current router structure.
    Some response bodies are modeled generically where the implementation is still evolving.
servers:
  - url: /api/v1
...
paths:
...
  /bootstrap:
    get:
      tags: [Bootstrap]
      operationId: getBootstrap
      summary: Get bootstrap metadata used by clients during app startup
      security:
        - bearerAuth: []
        - apiTokenAuth: []
        - {}
      responses:
        '200':
          description: Bootstrap payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Authenticate with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshTokens
      summary: Refresh access and refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Refresh succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/forgot-password:
    post:
      tags: [Auth]
      operationId: forgotPassword
      summary: Request a password reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '202':
          description: Reset flow accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectResponse'
  /auth/reset-password:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: Reset a password using a reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectResponse'
        '400':
          $ref: '#/components/responses/Error'
...
  /auth/me:
    get:
      tags: [Auth]
      operationId: me
      summary: Get the authenticated principal
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated user summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrincipalResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/api-tokens:
    get:
      tags: [API Tokens]
      operationId: listApiTokens
      summary: List API tokens
      security:
        - bearerAuth: []
      responses:
        '200':
          description: API token list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [API Tokens]
      operationId: createApiToken
      summary: Create an API token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiTokenCreateRequest'
      responses:
        '201':
          description: API token created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/api-tokens/{id}/revoke:
    post:
      tags: [API Tokens]
      operationId: revokeApiToken
      summary: Revoke an API token
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: API token revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Users]
      operationId: createUser
      summary: Create a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
      responses:
        '201':
          description: User created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/{id}:
    get:
      tags: [Users]
      operationId: getUser
      summary: Get a user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: User details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Users]
      operationId: putUser
      summary: Replace a user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
      responses:
        '200':
          description: User updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectResponse'
    patch:
      tags: [Users]
      operationId: patchUser
      summary: Partially update a user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
      responses:
        '200':
          description: User updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectResponse'
  /users/{id}/reset-password:
    post:
      tags: [Users]
      operationId: adminResetPassword
      summary: Reset a user password as an administrator
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
      responses:
        '200':
          description: Password reset succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectResponse'
  /audit:
    get:
      tags: [Audit]
      operationId: listAuditLogs
      summary: List audit log entries
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Audit log list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
...
  /settings/login-branding:
    get:
      tags: [Settings]
      operationId: getLoginBranding
      summary: Get editable login branding
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Login branding settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectResponse'
    put:
      tags: [Settings]
      operationId: updateLoginBranding
      summary: Update login branding
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
      responses:
        '200':
          description: Login branding updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectResponse'
  /settings/module-enablement:
    get:
      tags: [Module Enablement]
      operationId: getRuntimeModuleEnablement
      summary: Get runtime module enablement overrides
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Effective module enablement view.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleEnablementResponse'
    put:
      tags: [Module Enablement]
      operationId: updateRuntimeModuleEnablement
      summary: Update runtime module enablement overrides
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleEnablementUpdateRequest'
      responses:
        '200':
          description: Runtime overrides updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleEnablementResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiTokenAuth:
      type: apiKey
      in: header
      name: X-API-Token
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Numeric resource identifier
      schema:
        type: integer
        format: int64
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
    GenericObjectResponse:
      type: object
      additionalProperties: true
...
    VersionResponse:
      type: object
      required: [version, commit, buildDate]
      properties:
        version:
          type: string
        commit:
          type: string
        buildDate:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    ForgotPasswordRequest:
      type: object
      required: [usernameOrEmail]
      properties:
        usernameOrEmail:
          type: string
    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          format: password
    AuthTokenResponse:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
    PrincipalSummary:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
    PrincipalResponse:
      type: object
      additionalProperties: true
      properties:
        user:
          $ref: '#/components/schemas/PrincipalSummary'
    BootstrapResponse:
      type: object
      additionalProperties: true
      properties:
        version:
          type: integer
        generatedAt:
          type: string
          format: date-time
        app:
          $ref: '#/components/schemas/VersionResponse'
        branding:
          type: object
          additionalProperties: true
        modules:
          type: array
          items:
            type: object
            additionalProperties: true
        capabilities:
          type: object
          additionalProperties: true
        cache:
          type: object
          additionalProperties: true
        principal:
          $ref: '#/components/schemas/PrincipalSummary'
    ModuleEnablementUpdateRequest:
      type: object
      additionalProperties:
        type: boolean
    ModuleEnablementResponse:
      type: object
      additionalProperties: true
      properties:
        modules:
          type: array
          items:
            type: object
            additionalProperties: true
        flags:
          type: object
          additionalProperties:
            type: boolean
...
    ApiTokenCreateRequest:
      type: object
      additionalProperties: true
      properties:
        name:
          type: string
        permissions:
          type: array
          items:
            type: string
        expiresAt:
          type: string
          format: date-time
...
    PaginationMeta:
      type: object
      additionalProperties: true
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    ListEnvelope:
      type: object
      additionalProperties: true
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ErrorResponse:
      type: object
      additionalProperties: true
      properties:
        error:
          type: object
          additionalProperties: true
          properties:
            code:
              type: string
            message:
              type: string